+++ example/lib/extensions/payment_mode_ext.dart
@@ -4,7 +4,7 @@
 
 
 
-import 'package:hyperpay_plugin/flutter_hyperpay.dart';
+import 'package:hyperpay_payment_sdk/flutter_hyperpay.dart';
 
 /// A handy util to turn mode enums into their corresponding
 /// HyperPay string representation.