+++ lib/src/custom_ui/method_channel_custom_ui.dart
@@ -47,7 +47,7 @@ Future<PaymentResultData> implementPaymentCustomUI({
   } on PlatformException catch (e) {
     transactionStatus = "${e.message}";
     return PaymentResultData(
-        errorString: e.message, paymentResult: PaymentResult.error);
+        errorString: e.message, errorCode: e.code, paymentResult: PaymentResult.error);
   }
 }