+++ lib/src/custom_ui/method_channel_custom_ui_stc.dart
@@ -31,7 +31,7 @@ Future<PaymentResultData> implementPaymentCustomUISTC({
   } on PlatformException catch (e) {
     transactionStatus = "${e.message}";
     return PaymentResultData(
-        errorString: e.message, paymentResult: PaymentResult.error);
+        errorString: e.message, errorCode: e.code, paymentResult: PaymentResult.error);
   }
 }