+++ lib/src/ready_ui/method_channel_ready_ui.dart
@@ -44,7 +44,7 @@ Future<PaymentResultData> implementPayment(
   } on PlatformException catch (e) {
     transactionStatus = "${e.message}";
     return PaymentResultData(
-        errorString: e.message, paymentResult: PaymentResult.error);
+        errorString: e.message, errorCode: e.code, paymentResult: PaymentResult.error);
   }
 }